team.service.tsのvalidateProfileIdsをprofile.service.tsに移動 #6
