+++ src/team/team.service.ts
@@ -4,7 +4,6 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Profile } from 'src/entities/profile.entity';
 import { Team } from 'src/entities/team.entity';
 import { ProfileService } from 'src/profile/profile.service';
 import { Repository } from 'typeorm';
@@ -64,10 +63,13 @@ export class TeamService {
       throw new Error('Team needs at least one member');
     }
     const createdTeam = this.teamRepository.create(teamProps);
-    const members = await this.validateProfileIds(memberIds);
+    const members = await this.profileService.validateProfileIds(memberIds);
     createdTeam.members = members;
 
-    const invitees = await this.validateProfileIds(inviteeIds, members);
+    const invitees = await this.profileService.validateProfileIds(
+      inviteeIds,
+      members,
+    );
     createdTeam.invitees = invitees;
     return await this.teamRepository.save(createdTeam);
   }
@@ -81,10 +83,13 @@ export class TeamService {
     teamName && (updatedTeam.teamName = teamName);
     description && (updatedTeam.description = description);
 
-    const members = await this.validateProfileIds(memberIds);
+    const members = await this.profileService.validateProfileIds(memberIds);
     updatedTeam.members = members;
 
-    const invitees = await this.validateProfileIds(inviteeIds, members);
+    const invitees = await this.profileService.validateProfileIds(
+      inviteeIds,
+      members,
+    );
     updatedTeam.invitees = invitees;
     return await this.teamRepository.save(updatedTeam);
   }
@@ -97,41 +102,9 @@ export class TeamService {
     return team;
   }
 
-  async validateProfileIds(profileIds: string[], actualProfiles?: Profile[]) {
-    let profileIdsToRegister = [...profileIds];
-
-    // profileIdsのprofileが既にactualProfilesに存在しているか検証
-    if (actualProfiles) {
-      const removeProfileIds: string[] = [];
-      for (const profile of actualProfiles) {
-        if (profileIds.includes(profile.id)) {
-          removeProfileIds.push(profile.id);
-        }
-      }
-      // 存在する場合は削除
-      profileIdsToRegister = profileIds.filter(
-        (profileId) => !removeProfileIds.includes(profileId),
-      );
-    }
-
-    if (!profileIdsToRegister.length) {
-      return [];
-    }
-
-    const profiles = await this.profileService
-      .findByIds(profileIdsToRegister)
-      .catch((err) => {
-        if (err.status === 404) {
-          throw new BadRequestException('Invalid profileIds');
-        }
-        throw err;
-      });
-    return profiles;
-  }
-
   async addInvitees(id: string, profileIds: string[]) {
     const { invitees: currentInvitees } = await this.validateTeam(id);
-    const newInvitees = await this.validateProfileIds(
+    const newInvitees = await this.profileService.validateProfileIds(
       profileIds,
       currentInvitees,
     );
@@ -151,7 +124,7 @@ export class TeamService {
 
   async addMembers(id: string, profileIds: string[]) {
     const { members: currentMembers } = await this.validateTeam(id);
-    const newMembers = await this.validateProfileIds(
+    const newMembers = await this.profileService.validateProfileIds(
       profileIds,
       currentMembers,
     );
